Added lots of comments, fixed a few minor bugs
